requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
deep-translator==1.11.4
diff-match-patch==20241021
Django==5.2.1
django-import-export==4.3.7
idna==3.10
pillow==11.2.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
requests==2.32.3
soupsieve==2.7
sqlparse==0.5.3
tablib==3.8.0
typing_extensions==4.13.2
urllib3==2.4.0
